apply search and cateogry selection filtering

+++ app/src/main/java/com/sako/foodics_android_task/data/db/dao/CategoryDao.kt
@@ -24,7 +24,7 @@ interface CategoryDao {
     @Transaction
     @Query("SELECT * FROM localcategory WHERE id = :categoryId")
     suspend fun getCategoryWithProducts(
-        categoryId: Long
+        categoryId: Int
     ): LocalCategoryWithProducts?
